+++ CHANGELOG.md
@@ -19,8 +19,7 @@ nav_order: 1
   case `additional_disk_space` cannot be used.
 
 - Change `aiven_opensearch` field `opensearch_user_config.opensearch_version` (enum): add `3.3`
-- Add `aiven_organization_billing_group` field `payment_method`: Payment method.
-- Change `aiven_organization_billing_group` resource field `currency` (enum): optional ~~`true`~~ → `false`
+- Remove `aiven_organization_billing_group` field `currency`: field removed from resource and data source
 
 ## [4.50.0] - 2026-01-27